requirements.txt
flask==2.3.3
werkzeug==2.3.7
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.1.0
h5py==3.10.0
opencv-python==4.8.1.78
torch==2.1.2
torchvision==0.16.2
matplotlib==3.8.2
jupyter==1.0.0
notebook==7.0.6
scikit-learn==1.3.2
streamlit==1.29.0
mediapipe==0.10.8 
seaborn==0.12.2